requirements.txt
selenium>=4.15.0
python-dotenv>=1.0.0
faker>=20.1.0
webdriver-manager>=4.0.1
